chore(origin-ui-core): make parent component for asks and bids
